PassiveSkill追加

+++ KigurumiBreaker/Assets/Script/Player/SkillData/PassiveSkillDataList.cs
@@ -15,30 +15,30 @@ public class PassiveSkillData
     [Header("最大体力増加値")]
     [SerializeField] private int MaxHpAddNum;
     [Header("通常攻撃のダメージ上昇量(%)")]
-    [SerializeField] private int LowAttackDamageAddRate;
+    [SerializeField] private float LowAttackDamageAddRate;
     [Header("溜め攻撃のダメージ上昇量(%)")]
-    [SerializeField] private int ChargeAttackDamageAddRate;
+    [SerializeField] private float ChargeAttackDamageAddRate;
     [Header("遠距離攻撃のダメージ上昇量(%)")]
-    [SerializeField] private int RangedAttackDamageAddRate;
+    [SerializeField] private float RangedAttackDamageAddRate;
     [Header("遠距離攻撃の弾数増加量(個数)")]
     [SerializeField] private int RangedAttackBulletAddNum;
     [Header("移動速度上昇量(%)")]
-    [SerializeField] private int MoveSpeedAddRate;
+    [SerializeField] private float MoveSpeedAddRate;
     [Header("ダッシュ回数増加量(回数)")]
     [SerializeField] private int DashCountAddNum;
     [Header("ダメージカット率上昇量(%)")]
-    [SerializeField] private int DamageCutRateAddRate;
+    [SerializeField] private float DamageCutRateAddRate;
     [Header("回避率上昇量(%)")]
-    [SerializeField] private int DodgeRateAddRate;
+    [SerializeField] private float DodgeRateAddRate;
 
     public string skillName => SkillName;
     public int maxHpAddNum => MaxHpAddNum;
-    public int lowAttackDamageAddRate => LowAttackDamageAddRate;
-    public int chargeAttackDamageAddRate => ChargeAttackDamageAddRate;
-    public int rangedAttackDamageAddRate => RangedAttackDamageAddRate;
+    public float lowAttackDamageAddRate => LowAttackDamageAddRate;
+    public float chargeAttackDamageAddRate => ChargeAttackDamageAddRate;
+    public float rangedAttackDamageAddRate => RangedAttackDamageAddRate;
     public int rangedAttackBulletAddNum => RangedAttackBulletAddNum;
-    public int moveSpeedAddRate => MoveSpeedAddRate;
+    public float moveSpeedAddRate => MoveSpeedAddRate;
     public int dashCountAddNum => DashCountAddNum;
-    public int damageCutRateAddRate => DamageCutRateAddRate;
-    public int dodgeRateAddRate => DodgeRateAddRate;
+    public float damageCutRateAddRate => DamageCutRateAddRate;
+    public float dodgeRateAddRate => DodgeRateAddRate;
 }
